[lab7] change thickness type

+++ lab7/Slides/Style/LineStyle/CompositeLineStyle/CompositeLineStyle.h
@@ -11,7 +11,7 @@ public:
 	std::optional<RGBAColor> GetColor() const override;
 	void SetColor(const RGBAColor) override;
 
-	std::optional<uint8_t> GetThickness() const override;
+	std::optional<unsigned int> GetThickness() const override;
 	void SetThickness(const Thickness) override;
 
 private:
